#16 Add Lexer#peek_no_cache_read
The new method does a peek by first resetting the cache (#rewind_peek) and then doing the peek. It does set a new cache on its own peek.

This is needed to deal with overlap in tokens, specifically here between Labels and any token listed in search order before Label that could possibly be given as a chr state name.

For example, suppose the lexer is at '123abc' and I peek for a semicolon to see if we're at the end of a block. If we're not then the peek returns false, but we save @next_token = Number: 123 on the lexer (since Number is before Label in the list of Tokens to check).

Now that Number has been saved as the @next_token, there is no way to tell if you started with `123abc` or `123 abc`. The first case should have been read as a Label, the second case should be two tokens, a Number then a Label (so you can't just pop Number then pop Label to handle the first case because if it's actually the second case then you've processed two tokens that should have been one).

The same issue arises if your string is for example 'n=2', which would parse as a value pair in the above scenario but if it's a chrstate name then it should actually be parsed as a Label instead - and we crash because it wasn't (because it's neither a Label nor a Number, the two tokens we check for).

I wonder if there are other places in the code where something similar occurs and this could fix it.