#16 Add a DashLabel token for things like '-123abc' and use it
This fixes two issues:

1) Things like '123abc' should match as Labels, but things like '-123' used to have to match as Numbers because Labels couldn't start with '-', and things like '-123abc' would just cause a crash as chr state names because they didn't match Label or Number.

Having DashLabel, which matches non-quoted Labels that start with a dash, unifies the treatment of Labels and Numbers into Label and DashLabel, which I think more closely matches the original nexus description and Mesquite's behavior.

(Though Mesquite is more confusing because it sometimes treats '-' as a negative sign indicating a Number - but then reads the Number until the next punctuation, so would read -123abc-45gh -> '-123abc', '-45gh' - and sometimes treats '-' as a punctuation to be split on, as in, e.g. '-2e-3' -> '-', '2e', '-', '3', I'm not sure this behavior is consistent, and in any case we don't do any splitting on any punctuation, yet).

Of course if it would be okay to have Labels start with a dash then this could go away, but this seems reasonable to me and Label is used in several other places so I didn't want to mess with it.

This does make '-' special as a first character of a Label: why not other punctuation characters like '%' or '*'? I guess the answer here is because '-' is the only one that can be the first character of a number (not including '+') and we want to support numbers, positive and negative, as chr state names.

2) [Rewording the commit message of the previous commit] Using the new Lexer#peek_no_cache_read for peeking chr state names fixes the issue of overlap of Tokens, which I think is otherwise unavoidable.

The issue is that all Numbers that don't start with '-' overlap as Labels, all ValuePairs with no space and no quotes (and maybe more) overlap as Labels, all strings that start with 'ids' are MesquiteIds and also overlap as Labels, on and on.

Note that the case here applies to most of the tokens listed before Label - if a chr state name matches any of those non-Number tokens listed before Label then it could get peeked as non-Label/non-Number/non-DashLabel when peeking for a comma/semicolon to see if we're at the end of the block, and then we would crash when checking for the chr state name (because the token stored on the lexer for the lookaheaded chr state name is neither Label nor Number nor DashLabel).

I wrote tests for the Number and ValuePair cases since those seemed most likely to me, but again you could probably write a failing-prior-to-this-patch test for most tokens listed above Label in the search list.

The only place we use Number now is for character numbering, which should really only match positive integers, which we could easily do, but we may need to resurrect Number at some future point, so just leave it for now.