#16 "Fix" parsing of scientific notation Numbers
The previous Number#value was truncating the E... part of the notation on converting it to an integer.

I think it makes sense to leave in scientific notation since that's used to make numbers more recognizable and understandable: 1e-9 seems better than 0.000000009.

Note, for the chr state name case, '3e-2' is read as a Label, so stays that way; '-3e-2' is read as a Number (since a Label can't start with '-') so is now returned in the same way, as '-3e-2' (before it would have been -3, dropping the e part).

The other use case for Number is as the '1' in character '1 abc / def ghi' - there we probably shouldn't be allowing sci notation at all, or even negative numbers - as it stands, that Number gets to_i'ed, so truncates any e part.

Mesquite is a little confused here, at least to my eyes - in general I think Mesquite reads /-\n/ as the start of a number, but then it reads that number until it hits a punctuation character (of which '-' is included). So '-33e-260' would become two chr state names, '-33e' and '-260' (I hesitate to say, but this might be a bug - I think the intention would be to do '-33', 'e', '-260'). But when the number only has one digit there seems to be different behavior, so '-3e-2' becomes '-', '3e', '-', '2', i.e. in this case it seems to be treating '-' as a punctuation character and not the start of a negative number. And the 'e' character is not significant here, it could be any alpha string.

Compare https://github.com/MesquiteProject/MesquiteCore/blob/72ac14f5c3df2e58c20037e2b4cb4ae9dd474f36/Source/mesquite/lib/Parser.java#L937 (which I'm not sure is what's being used for these cases).