#16 Add a new CharacterLabel token allowing more than Label
(Unquoted) Label has to start with \w (and DashLabel had to start with -\w), CharacterLabel is much more permissive - that matches more closely what's allowed by TW and Mesquite, which as far as I can in both cases is arbitrary strings (maybe stripped).

Note we intentionally don't add CharacterLabel to the Tokens search list: we peek for it in the places we want to find it and don't want it searched for anywhere else.