#16 Replace Number with PositiveInteger
The only place Number was used was for number Characters, where it should really be a positive integer (I think).

I was previously hesitant to do this because Number may at some later time be needed again and then we either need to switch back to Number exclusively or have overlap with Number and PositiveInteger (not good).

But if you don't add PositiveInteger to the Tokens search list and you pop every time you peek a PositiveInteger then you've completely hidden its existence from the rest of the code, meaning you could resurrect Number independently.

If you were to peek a PositiveInteger and then not pop it for some reason then you could leak a PositiveInteger on the lexer into other code where you might have to do a peek_no_cache_read to get rid of it, the same as with any of multiple other token overlaps...